BackEnd actualizado a la base de datos

+++ Backend/src/main/java/com/apiweb/backend/Model/CalificacionesServicios.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class CalificacionesServicios {
     @Id
     private ObjectId calificador;
-    private int calificacion;
+    private double calificacion;
     private String comentario;
     private Instant fecha;
 }
